pyproject.toml
[project]
name = "deep-research-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.2.8",
    "langchain-openai>=1.1.0",
    "python-dotenv>=1.2.1",
]
